backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
redis==5.0.1
rq==1.15.1
PyPDF2==3.0.1
pdfplumber==0.10.3
pdf2image==1.16.3
pytesseract==0.3.10
Werkzeug==2.3.7
gunicorn==21.2.0
python-multipart==0.0.6